Complete rewrite.
Integrate the functionality of frakvol and frakclean.  Choose the operating
mode based on whether a path or a volume was given on the command line, and
fix a long-standing bug in frakvol where two volumes could not be fraked at
the same time in the same directory.

Add multiple options to better control frak's behavior and provide short
aliases for options that were already supported (diff lines, mount point
crossing, whether and where to write log files, munging of paths in the
output, debug mode, change directory creation, suppression of output to
standard out, suppression of diffs).

Fix a long-standing bug where changes to the directory ACLs or mode at the
top level of the volumes being compared weren't detected.

Refactor the code considerably, structuring the implementation much better
and eliminating large amounts of duplicate code and difficult-to-follow
nested if/then statements.  As part of that refactoring process, take
advantage of a few places where checks could be more efficient and
additional calls to stat could be avoided.

Added complete documentation.