fsr: Improve error handling for non-directory arguments
Before processing the files given on the command line, check each one to
see if it's a directory.  Ignore symlinks, non-existent directories, and
files that are not directories, warning about each one.  Previously, such
files would be silently ignored, leading fsr to exit successfully if the
only argument were a non-existent directory.  Suggested by Arthur
Prokosch.

Also add a pointer to up / afs-up in the man page in the discussion of fs
copyacl.