Disable loadmtpt in volcreate by default
volcreate no longer runs loadmtpt by default.  To enable loadmtpt
support, set $LOADMTPT in the configuration file to the path to the
loadmtpt utility.  Checking of the mount point prefix is only done if
loadmtpt support is enabled, since otherwise it doesn't matter how the
mount point path is designated.